refactor : State 클래스를 state 패키지 밑으로 이동
- 각 State 클래스의 이름을 다시 짓고 재사용하기 위해 패키지를 이동함
